+++ kotlin/test_kotlin.sh
@@ -13,9 +13,7 @@ YELLOW='\033[0;33m'
 NC='\033[0m' # No Color
 
 ROOT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")/.." && pwd)"
-
-TEST_RESULTS_DIR="$ROOT_DIR/kotlin/walletkit-tests/build/test-results/test"
-rm -rf "$TEST_RESULTS_DIR"
+KOTLIN_DIR="$ROOT_DIR/kotlin"
 
 cd "$ROOT_DIR"
 
@@ -45,7 +43,10 @@ echo -e "${BLUE}🔨 Step 1: Building Kotlin bindings with build_kotlin.sh${NC}"
 echo -e "${GREEN}✅ Kotlin bindings built${NC}"
 
 echo -e "${BLUE}📦 Step 2: Setting up Gradle test environment${NC}"
-cd "$ROOT_DIR/kotlin"
+cd "$KOTLIN_DIR"
+
+TEST_RESULTS_DIR="$(pwd -P)/walletkit-tests/build/test-results/test"
+rm -rf "$TEST_RESULTS_DIR"
 
 # Generate Gradle wrapper if missing
 if [ ! -f "gradlew" ]; then
@@ -76,7 +77,9 @@ echo ""
 echo -e "${BLUE}🧪 Step 3: Running Kotlin tests with verbose output...${NC}"
 echo ""
 
-./gradlew --no-daemon walletkit-tests:test --info --continue
+# Avoid reusing configuration cache entries from a different checkout, which can
+# redirect build outputs away from the current workspace and hide successful runs.
+./gradlew --no-daemon --no-configuration-cache walletkit-tests:test --info --continue
 
 echo ""
 echo "📊 Test Results Summary:"
