+++ kotlin/walletkit-tests/src/test/kotlin/org/world/walletkit/logging/LoggingTests.kt
@@ -26,7 +26,7 @@ private class CapturingLogger : Logger {
         }
 }
 
-class SimpleTest {
+class LoggingTests {
     @Test
     fun initLoggingForwardsLevelAndMessage() {
         val logger = CapturingLogger()
