+++ swift/tests/WalletKitTests/Logging/LoggingTests.swift
@@ -18,7 +18,7 @@ private final class CapturingLogger: WalletKit.Logger {
     }
 }
 
-final class SimpleTest: XCTestCase {
+final class LoggingTests: XCTestCase {
     func testInitLoggingForwardsLevelAndMessage() {
         let logger = CapturingLogger()
         WalletKit.initLogging(logger: logger, level: .info)
