--- swift/tests/WalletKitTests/TestHelpers.swift
@@ -1,40 +1,0 @@
-import Foundation
-import Security
-@testable import WalletKit
-
-func makeTempDirectory() -> URL {
-    let url = FileManager.default.temporaryDirectory.appendingPathComponent(
-        "walletkit-tests-\(UUID().uuidString)",
-        isDirectory: true
-    )
-    try? FileManager.default.createDirectory(at: url, withIntermediateDirectories: true)
-    return url
-}
-
-func uniqueKeystoreService() -> String {
-    "walletkit.devicekeystore.test.\(UUID().uuidString)"
-}
-
-func deleteKeychainItem(service: String, account: String) {
-    let query: [String: Any] = [
-        kSecClass as String: kSecClassGenericPassword,
-        kSecAttrService as String: service,
-        kSecAttrAccount as String: account
-    ]
-    SecItemDelete(query as CFDictionary)
-}
-
-func sampleCredential(
-    issuerSchemaId: UInt64 = 7,
-    expiresAt: UInt64 = 1_800_000_000
-) throws -> Credential {
-    let sampleCredentialJSON = """
-    {"id":13758530325042616850,"version":"V1","issuer_schema_id":\(issuerSchemaId),"sub":"0x114edc9e30c245ac8e1f98375f71668a9cd4e9f1e3e9b3385a1801e9d43d731b","genesis_issued_at":1700000000,"expires_at":\(expiresAt),"claims":["0x0000000000000000000000000000000000000000000000000000000000000000","0x0000000000000000000000000000000000000000000000000000000000000000","0x0000000000000000000000000000000000000000000000000000000000000000","0x0000000000000000000000000000000000000000000000000000000000000000","0x0000000000000000000000000000000000000000000000000000000000000000","0x0000000000000000000000000000000000000000000000000000000000000000","0x0000000000000000000000000000000000000000000000000000000000000000","0x0000000000000000000000000000000000000000000000000000000000000000","0x0000000000000000000000000000000000000000000000000000000000000000","0x0000000000000000000000000000000000000000000000000000000000000000","0x0000000000000000000000000000000000000000000000000000000000000000","0x0000000000000000000000000000000000000000000000000000000000000000","0x0000000000000000000000000000000000000000000000000000000000000000","0x0000000000000000000000000000000000000000000000000000000000000000","0x0000000000000000000000000000000000000000000000000000000000000000","0x0000000000000000000000000000000000000000000000000000000000000000"],"associated_data_hash":"0x0000000000000000000000000000000000000000000000000000000000000000","signature":null,"issuer":"0100000000000000000000000000000000000000000000000000000000000000"}
-    """
-    let bytes = Data(sampleCredentialJSON.utf8)
-    return try Credential.fromBytes(bytes: bytes)
-}
-
-func sampleBlindingFactor() -> FieldElement {
-    FieldElement.fromU64(value: 17)
-}
